Move error pages to logical place, add comments for index.js and App.js
