+++ src/App.js
@@ -6,12 +6,12 @@ import CssBaseline from "@material-ui/core/CssBaseline"
 import { makeStyles } from "@material-ui/core/styles"
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom"
 
+import Page404 from "components/ErrorPages/Page404"
+import Page500 from "components/ErrorPages/Page500"
 import Home from "components/home"
 import Login from "components/login"
 import Nav from "components/nav"
 import NewDraftWizard from "components/NewDraftWizard/NewDraftWizard"
-import Page404 from "error_pages/Page404"
-import Page500 from "error_pages/Page500"
 
 const useStyles = makeStyles(theme => ({
   content: {
@@ -35,6 +35,9 @@ const useStyles = makeStyles(theme => ({
   },
 }))
 
+/**
+ * Set up React router and app structure
+ */
 const App = () => {
   const classes = useStyles()
   return (
